+++ mediapipe/render/core/LUTFilter.cpp
@@ -3,31 +3,31 @@
 namespace Opipe
 {
     const std::string kLookupFragmentShaderString = SHADER_STRING(
-        varying highp vec2 vTexCoord;
-        varying highp vec2 vTexCoord1; // TODO: This is not used
+        varying lowp vec2 vTexCoord;
+        varying lowp vec2 vTexCoord1; // TODO: This is not used
 
         uniform sampler2D colorMap;
         uniform sampler2D colorMap1; // lookup texture
         uniform lowp float step;
 
         void main() {
-            highp vec4 textureColor = texture2D(colorMap, vTexCoord);
+            lowp vec4 textureColor = texture2D(colorMap, vTexCoord);
 
-            highp float blueColor = textureColor.b * 63.0;
+            lowp float blueColor = textureColor.b * 63.0;
 
-            highp vec2 quad1;
+            lowp vec2 quad1;
             quad1.y = floor(floor(blueColor) / 8.0);
             quad1.x = floor(blueColor) - (quad1.y * 8.0);
 
-            highp vec2 quad2;
+            lowp vec2 quad2;
             quad2.y = floor(ceil(blueColor) / 8.0);
             quad2.x = ceil(blueColor) - (quad2.y * 8.0);
 
-            highp vec2 texPos1;
+            lowp vec2 texPos1;
             texPos1.x = (quad1.x * 0.125) + 0.5 / 512.0 + ((0.125 - 1.0 / 512.0) * textureColor.r);
             texPos1.y = (quad1.y * 0.125) + 0.5 / 512.0 + ((0.125 - 1.0 / 512.0) * textureColor.g);
 
-            highp vec2 texPos2;
+            lowp vec2 texPos2;
             texPos2.x = (quad2.x * 0.125) + 0.5 / 512.0 + ((0.125 - 1.0 / 512.0) * textureColor.r);
             texPos2.y = (quad2.y * 0.125) + 0.5 / 512.0 + ((0.125 - 1.0 / 512.0) * textureColor.g);
