+++ mediapipe/render/core/OlaYUVTexture420P.cpp
@@ -5,9 +5,9 @@ NS_GI_BEGIN
 #if defined(__APPLE__)
 const std::string kYUVTextureFragmentShaderString = SHADER_STRING
 (
- varying highp vec2 vTexCoord;
- varying highp vec2 vTexCoord1;
- varying highp vec2 vTexCoord2;
+ varying lowp vec2 vTexCoord;
+ varying lowp vec2 vTexCoord1;
+ varying lowp vec2 vTexCoord2;
  uniform sampler2D colorMap;
  uniform sampler2D colorMap1;
  uniform sampler2D colorMap2;
@@ -36,9 +36,9 @@ const std::string kYUVTextureFragmentShaderString = SHADER_STRING
 const std::string kYUVTextureFragmentShaderString = SHADER_STRING
 (
  precision mediump float;
- varying highp vec2 vTexCoord;
- varying highp vec2 vTexCoord1;
- varying highp vec2 vTexCoord2;
+ varying lowp vec2 vTexCoord;
+ varying lowp vec2 vTexCoord1;
+ varying lowp vec2 vTexCoord2;
  
  uniform sampler2D colorMap;
  uniform sampler2D colorMap1;
