+++ mediapipe/render/module/beauty/filters/UnSharpMaskFilter.cpp
@@ -3,11 +3,11 @@
 namespace Opipe {
     const std::string kUnsharpMaskFragmentShaderString = SHADER_STRING
     (
-     varying highp vec2 vTexCoord;
-     varying highp vec2 vTexCoord1;
+     varying lowp vec2 vTexCoord;
+     varying lowp vec2 vTexCoord1;
      uniform sampler2D colorMap;
      uniform sampler2D colorMap1;
-     uniform highp float intensity;
+     uniform lowp float intensity;
      void main()
      {
         lowp vec4 sharpImageColor = texture2D(colorMap, vTexCoord);
